pyproject.toml
[project]
name = "xopen"
version = "0.1.0"
description = ""
authors = [
    {name = "Torben Andersen",email = "muldvang@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-magic (>=0.4.27,<0.5.0)"
]

[tool.poetry]
packages = [{include = "xopen", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
